pyproject.toml
[project]
name = "drt"
version = "0.1.0"
description = ""
authors = [
    {name = "Judy",email = "mjchoi0831@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph (>=0.4.8,<0.5.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-openai (>=0.3.25,<0.4.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "ipython (>=9.3.0,<10.0.0)",
    "langchain-anthropic (>=0.3.15,<0.4.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
